+++ demosplan/DemosPlanCoreBundle/Controller/Survey/SurveyController.php
@@ -14,8 +14,8 @@ use demosplan\DemosPlanCoreBundle\Annotation\DplanPermissions;
 use demosplan\DemosPlanCoreBundle\Controller\Base\BaseController;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
 use demosplan\DemosPlanCoreBundle\Exception\MessageBagException;
+use demosplan\DemosPlanCoreBundle\Exception\SurveyInputDataException;
 use demosplan\DemosPlanProcedureBundle\Logic\ProcedureHandler;
-use demosplan\DemosPlanSurveyBundle\Exception\SurveyInputDataException;
 use demosplan\DemosPlanSurveyBundle\Logic\SurveyCreateHandler;
 use demosplan\DemosPlanSurveyBundle\Logic\SurveyHandler;
 use demosplan\DemosPlanSurveyBundle\Logic\SurveyNewHandler;
