+++ tests/backend/core/Survey/Unit/SurveyUpdateHandlerTest.php
@@ -11,7 +11,7 @@
 namespace Tests\Core\Survey\Unit;
 
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
-use demosplan\DemosPlanSurveyBundle\Exception\SurveyInputDataException;
+use demosplan\DemosPlanCoreBundle\Exception\SurveyInputDataException;
 use demosplan\DemosPlanSurveyBundle\Logic\SurveyUpdateHandler;
 use Exception;
